refactor: move ticket-history into screen-components
