+++ src/screen-components/ticket-history/TicketHistoryScreenComponent.tsx
@@ -7,10 +7,7 @@ import {
 import {useTimeContext} from '@atb/time';
 import {TicketingTexts, useTranslation} from '@atb/translations';
 import {View} from 'react-native';
-import {
-  TicketHistoryMode,
-  TicketHistoryScreenParams,
-} from '@atb/ticket-history';
+import {TicketHistoryMode, TicketHistoryScreenParams} from './types';
 import {TicketHistoryModeTexts} from '@atb/translations/screens/Ticketing';
 import {useAuthContext} from '@atb/auth';
 import React from 'react';
