+++ src/screen-components/ticket-history/index.tsx
@@ -1,0 +1,2 @@
+export {TicketHistoryScreenComponent} from './TicketHistoryScreenComponent';
+export type {TicketHistoryMode, TicketHistoryScreenParams} from './types';
