+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/navigation-types.ts
@@ -9,7 +9,7 @@ import {
 import {TicketTabNavStackParams} from './Ticketing_TicketTabNavStack/navigation-types';
 import {TabNavigatorScreenProps} from '../navigation-types';
 import {StackScreenProps} from '@react-navigation/stack';
-import {TicketHistoryScreenParams} from '@atb/ticket-history';
+import {TicketHistoryScreenParams} from '@atb/screen-components/ticket-history';
 
 export type TicketingStackParams = StackParams<{
   Ticketing_NotEnabledScreen: undefined;
