--- src/ticket-history/index.tsx
@@ -1,3 +1,0 @@
-export {TicketHistoryScreenComponent} from './TicketHistoryScreenComponent';
-export type {TicketHistoryMode} from './types';
-export type {TicketHistoryScreenParams} from './types';
